requirements.txt
streamlit
google-generativeai>=0.8.0
fpdf
beautifulsoup4
requests
dnspython
pandas